bhsi-backend/app/services/embedder/requirements.txt
# web framework
fastapi>=0.111.0,<0.112.0  # first GA release on Pydantic 2
uvicorn[standard]>=0.30.0,<0.31.0  # 0.30 works on Py≥3.9 and FastAPI≥0.110

# Gen AI SDK
google-generativeai==0.8.3  # Correct package name and stable version

# infra / helpers
tenacity==8.2.3
python-dotenv==1.0.1
anyio>=4.8.0,<5.0.0        # Required for async operations 